ParserUtil: Refactor accounts related into its own util class
