+++ src/main/java/seedu/address/logic/parser/ParserUtil.java
@@ -11,8 +11,6 @@ import java.util.stream.Stream;
 import seedu.address.commons.core.index.Index;
 import seedu.address.commons.util.StringUtil;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.accounts.Password;
-import seedu.address.model.accounts.Username;
 import seedu.address.model.ingredient.IngredientName;
 import seedu.address.model.ingredient.IngredientPrice;
 import seedu.address.model.ingredient.IngredientUnit;
@@ -189,36 +187,6 @@ public class ParserUtil {
             throw new ParseException(Price.MESSAGE_PRICE_CONSTRAINTS);
         }
         return new Price(trimmedPrice);
-    }
-
-    //================ Authentication Commands Parser Util ===================================================
-
-    /**
-     * Parses a {@code username} into a {@code Username}. Leading and trailing whitespaces will be trimmed.
-     *
-     * @throws ParseException if the given {@code username} is invalid.
-     */
-    public static Username parseUsername(String username) throws ParseException {
-        requireNonNull(username);
-        String trimmedUsername = username.trim();
-        if (!Username.isValidUsername(trimmedUsername)) {
-            throw new ParseException(Username.MESSAGE_USERNAME_CONSTRAINT);
-        }
-        return new Username(trimmedUsername);
-    }
-
-    /**
-     * Parses a {@code password} into a {@code Password}. Leading and trailing whitespaces will be trimmed.
-     *
-     * @throws ParseException if the given {@code password} is invalid.
-     */
-    public static Password parsePassword(String password) throws ParseException {
-        requireNonNull(password);
-        String trimmedPassword = password.trim();
-        if (!Password.isValidPassword(trimmedPassword)) {
-            throw new ParseException(Password.MESSAGE_PASSWORD_CONSTRAINT);
-        }
-        return new Password(trimmedPassword);
     }
 
     //================ Ingredients Parser Util ===================================================
