+++ src/main/java/seedu/address/logic/parser/accounts/RegisterCommandParser.java
@@ -10,7 +10,6 @@ import seedu.address.logic.commands.accounts.RegisterCommand;
 import seedu.address.logic.parser.ArgumentMultimap;
 import seedu.address.logic.parser.ArgumentTokenizer;
 import seedu.address.logic.parser.Parser;
-import seedu.address.logic.parser.ParserUtil;
 import seedu.address.logic.parser.Prefix;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.accounts.Account;
@@ -35,8 +34,8 @@ public class RegisterCommandParser implements Parser<RegisterCommand> {
             throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, RegisterCommand.MESSAGE_USAGE));
         }
 
-        Username username = ParserUtil.parseUsername(argMultimap.getValue(PREFIX_ID).get());
-        Password password = ParserUtil.parsePassword(argMultimap.getValue(PREFIX_PASSWORD).get());
+        Username username = AccountParserUtil.parseUsername(argMultimap.getValue(PREFIX_ID).get());
+        Password password = AccountParserUtil.parsePassword(argMultimap.getValue(PREFIX_PASSWORD).get());
 
         Account account = new Account(username, password);
