+++ src/main/java/seedu/address/model/Model.java
@@ -4,7 +4,6 @@ import java.util.Map;
 import java.util.function.Predicate;
 
 import javafx.collections.ObservableList;
-
 import seedu.address.logic.commands.menu.SortMenuCommand.SortMethod;
 import seedu.address.model.accounts.Account;
 import seedu.address.model.ingredient.Ingredient;
