+++ src/main/resources/view/BrowserPanel.fxml
@@ -2,7 +2,6 @@
 
 <?import javafx.scene.layout.StackPane?>
 <?import javafx.scene.web.WebView?>
-
 <StackPane xmlns:fx="http://javafx.com/fxml/1">
   <WebView fx:id="browser"/>
 </StackPane>
