+++ src/main/resources/view/IngredientListPanel.fxml
@@ -2,7 +2,6 @@
 
 <?import javafx.scene.control.ListView?>
 <?import javafx.scene.layout.VBox?>
-
 <VBox xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1">
     <ListView fx:id="ingredientListView" VBox.vgrow="ALWAYS" />
 </VBox>
