+++ src/main/resources/view/ReservationListPanel.fxml
@@ -2,7 +2,6 @@
 
 <?import javafx.scene.control.ListView?>
 <?import javafx.scene.layout.VBox?>
-
 <VBox xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1">
   <ListView fx:id="reservationListView" VBox.vgrow="ALWAYS" />
 </VBox>
