+++ src/main/resources/view/SalesReportWindow.fxml
@@ -3,11 +3,10 @@
 <?import javafx.scene.control.Label?>
 <?import javafx.scene.control.TableColumn?>
 <?import javafx.scene.control.TableView?>
-<?import javafx.scene.layout.VBox?>
-
-<?import javafx.scene.Scene?>
 <?import javafx.scene.image.Image?>
-
+<?import javafx.scene.layout.VBox?>
+<?import javafx.scene.Scene?>
+<?import javafx.stage.Stage?>
 <fx:root type="javafx.stage.Stage" xmlns="http://javafx.com/javafx/9.0.4" xmlns:fx="http://javafx.com/fxml/1">
   <icons>
     <Image url="@/images/sales.png" />
