+++ src/main/resources/view/StatusBarFooter.fxml
@@ -3,7 +3,6 @@
 <?import org.controlsfx.control.StatusBar?>
 <?import javafx.scene.layout.ColumnConstraints?>
 <?import javafx.scene.layout.GridPane?>
-
 <GridPane styleClass="grid-pane" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1">
   <columnConstraints>
     <ColumnConstraints hgrow="SOMETIMES" minWidth="10" prefWidth="100" />
