+++ src/test/java/guitests/guihandles/sales/RecordCardHandle.java
@@ -1,10 +1,8 @@
 package guitests.guihandles.sales;
 
 import guitests.guihandles.NodeHandle;
-
 import javafx.scene.Node;
 import javafx.scene.control.Label;
-
 import seedu.address.model.salesrecord.SalesRecord;
 
 /**
